switched to uint8 and 16

+++ quantum/process_keycode/process_tap_dance.c
@@ -132,9 +132,9 @@ bool process_tap_dance(uint16_t keycode, keyrecord_t *record) {
 void matrix_scan_tap_dance () {
   if (highest_td == -1)
     return;
-  int tap_user_defined;
+  uint16_t tap_user_defined;
 
-for (int i = 0; i <= highest_td; i++) {
+for (uint8_t i = 0; i <= highest_td; i++) {
     qk_tap_dance_action_t *action = &tap_dance_actions[i];
     if(action->custom_tapping_term > 0 ) {
       tap_user_defined = action->custom_tapping_term;
